\Refactor: Restrict tolerance comparator to painter tests only
- Removed global tolerance comparator from flutter_test_config.dart
- Moved _TolerantComparator to painter_test_helpers.dart
- Added setupPainterGoldenComparator() and tearDownPainterGoldenComparator()
- Updated all 4 painter test files to use setUp/tearDown for comparator
- Modified _TolerantComparator.compare to call super.compare() on failure
  to generate proper diff/failure artifacts for debugging

Benefits:
- Painter tests use 0.5% tolerance for cross-platform compatibility
- Other tests preserve strict golden file comparison
- Failure artifacts are generated when tolerance is exceeded
- Better isolation and explicit test configuration

All 2834 tests passing."
