+++ test/widget_tests/painters/telegram_logo_test.dart
@@ -6,6 +6,14 @@ import 'painter_test_helpers.dart';
 
 void main() {
   group('TelegramLogo Painter Tests', () {
+    setUp(() {
+      setupPainterGoldenComparator();
+    });
+
+    tearDown(() {
+      tearDownPainterGoldenComparator();
+    });
+
     testWidgets('should render TelegramLogo correctly (golden test)',
         (WidgetTester tester) async {
       const key = ValueKey('telegram_logo_painter');
